requirements.txt
Flask==2.3.2
flask-bootstrap==3.3.7.1
flask-sqlalchemy==3.0.5
python-dotenv==1.0.0
